font styles + rename
